deps: reinstate support for old versions of numpy? (#299)
* reinstate old numpy and modify ci to test

* use pip install -e in python installation directions in README
This updates the README to allow any version of cython and numpy.
When the user runs 'pip install -e', pip will automatically install the
appropriate versions of cython and numpy based on the version
constraints in pyproject.toml and setup.py.
Note that cython and numpy are build-time dependencies and only numpy
is a run-time dependency. This means that pip will temporarily install
numpy >2.0 (and cython) to compile pgenlib but they won't be installed
when the user imports pgenlib. In fact, numpy 1.19.3 (instead of 2.0)
could be installed when the user imports pgenlib.

* add py3.13 to release workflow